refactor(getScrollbarWidth): move to a separate util file
